+++ src/nrnmpi/nrnmpi_impl.h
@@ -3,8 +3,8 @@
 
 #include <mpi.h>
 
-extern void* hoc_Emalloc(unsigned long size);
-extern void* hoc_Erealloc(void* buf, unsigned long size);
+extern void* hoc_Emalloc(size_t size);
+extern void* hoc_Erealloc(void* buf, size_t size);
 extern void hoc_malchk();
 
 extern MPI_Comm nrnmpi_world_comm;
